refactor: 채점 서비스 모듈 분리
